ARM64 fixes.

+++ Dependencies/MKNumberBadgeView/MKNumberBadgeView.m
@@ -103,7 +103,7 @@
 	
 	CGContextRef curContext = UIGraphicsGetCurrentContext();
 
-	NSString* numberString = [NSString stringWithFormat:@"%d",self.value];
+	NSString* numberString = [NSString stringWithFormat:@"%lu", (unsigned long)self.value];
 	
 	
 	CGSize numberSize = [numberString sizeWithFont:self.font];
@@ -266,7 +266,7 @@
 
 - (CGSize)badgeSize
 {
-	NSString* numberString = [NSString stringWithFormat:@"%d",self.value];
+	NSString* numberString = [NSString stringWithFormat:@"%lu",(unsigned long)self.value];
 	
 	
 	CGSize numberSize = [numberString sizeWithFont:self.font];
